remove files that are (most likely) used nowhere
